+++ Tests/Editor/AssetGroupTests.cs
@@ -1,3 +1,4 @@
+using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 using System.IO;